requirements.txt
celery[redis]==5.3.1
Django==4.2.3
django-celery-results==2.5.1
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-celery-beat==2.5.0
drf-yasg==1.21.6
flower==2.0.0
psycopg2-binary==2.9.6
